requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
cytoolz==1.0.1
eth-account==0.13.6
eth-abi==5.2.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.4.0
eth-rlp==2.2.0
eth-typing==3.5.2
eth-utils==2.3.1
frozenlist==1.6.0
hexbytes==1.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
parsimonious==0.10.0
protobuf==6.30.2
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rlp==4.1.0
toolz==1.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
web3==6.10.0
websockets==15.0.1
yarl==1.20.0